Visibility has been implemented
Visibility has been added to the map rendering, and can be changed via the DEFAULT_VIS global constant in the Hero.h file.
It still needs testing when near the edges, but it's hard to test without movement implemented. That should definitely be high on the priority list!
I also changed a couple things structurally, but nothing worth mentioning since it's all so new anyways.